Project/sparkpredictor/requirements.txt
kafka-python==1.4.7
loguru==0.3.2
pandas==0.25.2
